+++ examples/bri-3/src/app.module.ts
@@ -1,7 +1,8 @@
 import { Module } from '@nestjs/common';
 import { IdentityModule } from './bri/identity/identity.module';
+import { WorkgroupModule } from './bri/workgroup/workgroup.module';
 
 @Module({
-  imports: [IdentityModule],
+  imports: [IdentityModule, WorkgroupModule],
 })
 export class AppModule {}
