+++ examples/bri-3/src/bri/workgroup/workflows/models/workflow.ts
@@ -1,5 +1,5 @@
 import { IWorkflow } from './workflow.interface';
-import { Workstep } from '../workstep/workstep';
+import { Workstep } from '../../worksteps/models/workstep';
 
 export class Workflow implements IWorkflow {
   id: string; // TODO: Add uuid after #491
