+++ examples/bri-3/src/bri/workgroup/workgroup.module.ts
@@ -1,0 +1,6 @@
+import { Module } from '@nestjs/common';
+
+@Module({
+  imports: []
+})
+export class WorkgroupModule {}
